flex label in separate file
